GH-19: store refactoring
