+++ frontend/src/store/WebRtc/mobilenet-ssd.json
